+++ crates/rstix/src/validate/checks/pattern_parse.rs
@@ -2,4 +2,5 @@
 
 use super::{ValidationContext, ValidationReport};
 
+#[allow(dead_code)]
 pub fn run(_ctx: &ValidationContext<'_>, _report: &mut ValidationReport) {}